claude-python-101/pyproject.toml
[project]
name = "claude-python-101"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic[bedrock]>=0.86.0",
    "ipykernel>=7.2.0",
    "python-dotenv>=1.2.2",
]
